Refocus ADR repo around specifications
After feedback that the journal approach of the typical ADR system
is not a good format for aiding development of clients we are
refocussing this repository around specifications and larger, living,
documents.

We will remain flexible in what goes in here - being useful is what
matters, the journal style did not work.

Future PRs might update old ADRs and combine them into later ones
marking the old ones as such.

Signed-off-by: R.I.Pienaar <rip@devco.net>